+++ www/go/core/views/extjs3/customfields/EditFieldSetDialog.js
@@ -11,6 +11,26 @@ go.customfields.EditFieldSetDialog = Ext.extend(go.form.Dialog, {
 		fs[0].title = false;
 		return items.concat(fs);
 	},
+
+	/**
+	 * override the getCustomFieldSets function as the original filters out the field sets that are rendered in their own
+	 * tabs. In this use case, a field set should always be returned.
+	 *
+	 * @returns {*[]}
+	 */
+	getCustomFieldSets : function() {
+		const items = [];
+		const fieldsets = go.customfields.CustomFields.getFormFieldSets(this.entityStore);
+		fieldsets.forEach(function(fs) {
+			if(fs.fieldSet.permissionLevel <= 10) {
+				return;
+			}
+			fs.columnWidth = 1;
+			items.push(fs);
+		}, this);
+
+		return items;
+	},
 });